Relocate provider services
The provider services are now split out into their own files. This
should make it easier to identify the location of each.
