+++ domain/cloud/service/package_test.go
@@ -6,6 +6,8 @@ package service
 import (
 	"testing"
 
+	jujutesting "github.com/juju/testing"
+	gomock "go.uber.org/mock/gomock"
 	gc "gopkg.in/check.v1"
 )
 
@@ -13,4 +15,18 @@ import (
 
 func TestPackage(t *testing.T) {
 	gc.TestingT(t)
+}
+
+type baseSuite struct {
+	jujutesting.IsolationSuite
+
+	state          *MockState
+	watcherFactory *MockWatcherFactory
+}
+
+func (s *baseSuite) setupMocks(c *gc.C) *gomock.Controller {
+	ctrl := gomock.NewController(c)
+	s.state = NewMockState(ctrl)
+	s.watcherFactory = NewMockWatcherFactory(ctrl)
+	return ctrl
 }
