+++ domain/cloud/service/service.go
@@ -22,24 +22,16 @@ type WatcherFactory interface {
 
 // State describes retrieval and persistence methods for storage.
 type State interface {
+	ProviderState
+
 	// UpsertCloud persists the input cloud entity.
 	UpsertCloud(context.Context, cloud.Cloud) error
 
 	// DeleteCloud deletes the input cloud entity.
 	DeleteCloud(context.Context, string) error
 
-	// Cloud returns the cloud with the specified name.
-	Cloud(context.Context, string) (*cloud.Cloud, error)
-
 	// ListClouds returns the clouds matching the optional filter.
 	ListClouds(context.Context) ([]cloud.Cloud, error)
-
-	// WatchCloud returns a new NotifyWatcher watching for changes to the specified cloud.
-	WatchCloud(
-		ctx context.Context,
-		getWatcher func(string, string, changestream.ChangeType) (watcher.NotifyWatcher, error),
-		name string,
-	) (watcher.NotifyWatcher, error)
 }
 
 // Service provides the API for working with clouds.
@@ -98,50 +90,5 @@ func NewWatchableService(st State, watcherFactory WatcherFactory) *WatchableServ
 
 // WatchCloud returns a watcher that observes changes to the specified cloud.
 func (s *WatchableService) WatchCloud(ctx context.Context, name string) (watcher.NotifyWatcher, error) {
-	return s.st.WatchCloud(ctx, s.watcherFactory.NewValueWatcher, name)
-}
-
-// ProviderService provides the API for working with clouds.
-// The provider service is a subset of the cloud service, and is used by the
-// provider package to interact with the cloud service. By not exposing the
-// full cloud service, the provider package is not able to modify the cloud
-// entities, only read them.
-type ProviderService struct {
-	st State
-}
-
-// NewProviderService returns a new service reference wrapping the input state.
-func NewProviderService(st State) *ProviderService {
-	return &ProviderService{
-		st: st,
-	}
-}
-
-// Cloud returns the named cloud.
-func (s *ProviderService) Cloud(ctx context.Context, name string) (*cloud.Cloud, error) {
-	cloud, err := s.st.Cloud(ctx, name)
-	return cloud, errors.Trace(err)
-}
-
-// WatchableProviderService returns a new provider service reference wrapping
-// the input state and watcher factory.
-type WatchableProviderService struct {
-	ProviderService
-	watcherFactory WatcherFactory
-}
-
-// NewWatchableProviderService returns a new service reference wrapping the
-// input state and watcher factory.
-func NewWatchableProviderService(st State, watcherFactory WatcherFactory) *WatchableProviderService {
-	return &WatchableProviderService{
-		ProviderService: ProviderService{
-			st: st,
-		},
-		watcherFactory: watcherFactory,
-	}
-}
-
-// WatchCloud returns a watcher that observes changes to the specified cloud.
-func (s *WatchableProviderService) WatchCloud(ctx context.Context, name string) (watcher.NotifyWatcher, error) {
 	return s.st.WatchCloud(ctx, s.watcherFactory.NewValueWatcher, name)
 }
