+++ domain/modelconfig/service/service.go
@@ -28,12 +28,7 @@ type ModelDefaultsProvider interface {
 // State represents the state entity for accessing and setting per
 // model configuration values.
 type State interface {
-	// AllKeysQuery returns a SQL statement that will return all known model config
-	// keys.
-	AllKeysQuery() string
-
-	// ModelConfig returns the currently set config for the model.
-	ModelConfig(context.Context) (map[string]string, error)
+	ProviderState
 
 	// ModelConfigHasAttributes returns the set of attributes that model config
 	// currently has set out of the list supplied.
@@ -331,61 +326,5 @@ func NewWatchableService(
 // Watch returns a watcher that returns keys for any changes to model
 // config.
 func (s *WatchableService) Watch() (watcher.StringsWatcher, error) {
-	return s.watcherFactory.NewNamespaceWatcher("model_config", changestream.All, s.st.AllKeysQuery())
-}
-
-// ProviderService defines the service for interacting with ModelConfig.
-// The provider service is a subset of the ModelConfig service, and is used by
-// the provider package to interact with the ModelConfig service. By not
-// exposing the full ModelConfig service, the provider package is not able to
-// modify the ModelConfig entities, only read them.
-type ProviderService struct {
-	st State
-}
-
-// NewProviderService creates a new ModelConfig service.
-func NewProviderService(
-	st State,
-) *ProviderService {
-	return &ProviderService{
-		st: st,
-	}
-}
-
-// ModelConfig returns the current config for the model.
-func (s *ProviderService) ModelConfig(ctx context.Context) (*config.Config, error) {
-	stConfig, err := s.st.ModelConfig(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("getting model config from state: %w", err)
-	}
-
-	altConfig := transform.Map(stConfig, func(k, v string) (string, any) { return k, v })
-	return config.New(config.NoDefaults, altConfig)
-}
-
-// WatchableProviderService defines the service for interacting with ModelConfig
-// and the ability to create watchers.
-type WatchableProviderService struct {
-	ProviderService
-	watcherFactory WatcherFactory
-}
-
-// NewWatchableProviderService creates a new WatchableProviderService for
-// interacting with ModelConfig and the ability to create watchers.
-func NewWatchableProviderService(
-	st State,
-	watcherFactory WatcherFactory,
-) *WatchableProviderService {
-	return &WatchableProviderService{
-		ProviderService: ProviderService{
-			st: st,
-		},
-		watcherFactory: watcherFactory,
-	}
-}
-
-// Watch returns a watcher that returns keys for any changes to model
-// config.
-func (s *WatchableProviderService) Watch() (watcher.StringsWatcher, error) {
 	return s.watcherFactory.NewNamespaceWatcher("model_config", changestream.All, s.st.AllKeysQuery())
 }
